+++ packages/odata-service-writer/src/types.ts
@@ -1,3 +1,4 @@
+import { ProxyBackend } from '@sap-ux/ui5-config';
 export enum OdataVersion {
     v2 = '2',
     v4 = '4'
@@ -18,4 +19,5 @@ export interface OdataService {
         technicalName: string;
         xml: string;
     };
+    previewSettings?: Partial<ProxyBackend>;
 }