+++ packages/ui5-application-writer/src/data/defaults.ts
@@ -42,6 +42,15 @@ export function appDefaults(app: App): App {
     };
 }
 
+export const UI5_DEFAULT = {
+    DEFAULT_UI5_VERSION: '',
+    DEFAULT_LOCAL_UI5_VERSION: '1.95.0',
+    MIN_LOCAL_SAPUI5_VERSION: '1.76.0',
+    MIN_LOCAL_OPENUI5_VERSION: '1.52.5',
+    SAPUI5_CDN: 'https://ui5.sap.com',
+    OPENUI5_CDN: 'https://openui5.hana.ondemand.com'
+};
+
 /**
  * Merges version properties with the provided UI5 instance.
  *
@@ -51,11 +60,32 @@ export function appDefaults(app: App): App {
 export function mergeUi5(ui5?: UI5): UI5 {
     const merged: Partial<UI5> = {
         minUI5Version: ui5?.minUI5Version || '1.60',
-        version: ui5?.version || '1.84.0'
+        version: ui5?.version || UI5_DEFAULT.DEFAULT_UI5_VERSION // no version indicates the latest available should be used
     };
+    merged.framework = ui5?.framework || 'SAPUI5';
+    merged.frameworkUrl =
+        ui5?.frameworkUrl || merged.framework === 'SAPUI5' ? UI5_DEFAULT.SAPUI5_CDN : UI5_DEFAULT.OPENUI5_CDN;
+
+    // if a specific local version is provided, use it, otherwise, sync with version but keep minimum versions in mind
+    if (ui5?.localVersion) {
+        merged.localVersion = ui5.localVersion;
+    } else {
+        merged.localVersion =
+            // eslint-disable-next-line no-nested-ternary
+            merged.version === UI5_DEFAULT.DEFAULT_UI5_VERSION
+                ? UI5_DEFAULT.DEFAULT_LOCAL_UI5_VERSION
+                : merged.framework === 'SAPUI5'
+                    ? UI5_DEFAULT.MIN_LOCAL_SAPUI5_VERSION
+                    : UI5_DEFAULT.MIN_LOCAL_OPENUI5_VERSION; // minimum version available as local libs
+        if (parseFloat(merged.version!) > parseFloat(merged.localVersion)) {
+            merged.localVersion = merged.version!;
+        }
+    }
+
     merged.descriptorVersion =
         ui5?.descriptorVersion || (mappings as Record<string, string>)[merged.minUI5Version!] || '1.12.0';
-    merged.typesVersion = ui5?.typesVersion || parseFloat(merged.version!) >= 1.76 ? merged.version : '1.71.18';
+    merged.typesVersion =
+        ui5?.typesVersion || parseFloat(merged.localVersion!) >= 1.76 ? merged.localVersion : '1.71.18';
     merged.ui5Theme = ui5?.ui5Theme || 'sap_fiori_3';
     // Return merged, does not update passed ref
     return Object.assign({}, ui5, merged) as UI5;