+++ packages/ui5-application-writer/src/data/index.ts
@@ -1,5 +1,6 @@
 import { Ui5App } from '../types';
 import { appDefaults, packageDefaults, mergeUi5 } from './defaults';
+import { validate } from './validators';
 
 /**
  * Merges Ui5App instance with default properties.
@@ -9,6 +10,7 @@ import { appDefaults, packageDefaults, mergeUi5 } from './defaults';
  * @returns {Ui5App} - the updated Ui5App instance
  */
 export function mergeWithDefaults(ui5App: Ui5App): Ui5App {
+    validate(ui5App);
     ui5App.app = appDefaults(ui5App.app);
     ui5App.ui5 = mergeUi5(ui5App.ui5);
     ui5App.package = Object.assign(packageDefaults(ui5App.package.version, ui5App.app.description), ui5App.package);