+++ packages/ui5-config/src/types.ts
@@ -1,37 +1,53 @@
-export interface Destination {
-    destination: string;
-    destinationInstance?: string;
-}
-export type Backend = Partial<Destination> & {
-    path: string;
+export interface ProxyBackend {
+    path?: string;
     url: string;
-};
+    client?: string;
+    destination?: string;
+    destinationInstance?: string;
+    pathPrefix?: string;
+    scp?: boolean;
+    apiHub?: boolean;
+    ws?: boolean;
+    xfwd?: boolean;
+}
 
-export interface MiddlewareConfig {
+export interface ProxyUIConfig {
+    [key: string]: unknown | undefined;
+    path?: string[];
+    url?: string;
+    version?: string;
+    directLoad?: boolean;
+}
+
+export interface CustomMiddleware<C extends any> {
     name: string;
     beforeMiddleware?: string;
     afterMiddleware?: string;
     mountPath?: string;
-    configuration: {
-        port?: number;
-        path?: string;
-        annotations?: {
-            localPath?: string;
-            urlPath: string;
-        };
-        service?: {
-            urlBasePath: string;
-            name: string;
-            metadataXmlPath: string;
-            mockdataRootPath?: string;
-            generateMockData?: boolean;
-        };
-        backend?: Backend[];
-        ui5?: {
-            path: string[];
-            url: string;
-            version: string | null;
-        };
-        ignoreCertError?: boolean;
+    configuration: C;
+}
+
+export interface FioriAppReloadConfig {
+    port: number;
+    path: string;
+}
+
+export interface FioriToolsProxyConfig {
+    backend?: ProxyBackend[];
+    ui5?: ProxyUIConfig;
+    ignoreCertError?: boolean;
+}
+
+export interface MockserverConfig {
+    annotations?: {
+        localPath?: string;
+        urlPath: string;
+    };
+    service?: {
+        urlBasePath: string;
+        name: string;
+        metadataXmlPath: string;
+        mockdataRootPath?: string;
+        generateMockData?: boolean;
     };
 }